+++ src/gcovr/formats/gcov/parser/text.py
@@ -48,6 +48,7 @@ from typing import (
 
 from ....data_model.coverage import FileCoverage, LineCoverage
 from ....data_model.merging import FUNCTION_MAX_LINE_MERGE_OPTIONS, MergeOptions
+from ....exceptions import SanityCheckError
 from ....logging import LOGGER
 from ....utils import get_md5_hexdigest
 
@@ -55,6 +56,8 @@ from .common import (
     SUSPICIOUS_COUNTER,
     check_hits,
 )
+
+UNKNOWN_FUNCTION_NAME = "<unknown function>"
 
 
 def _line_pattern(pattern: str) -> Pattern[str]:
@@ -159,8 +162,8 @@ class _FunctionLine(NamedTuple):
     """
 
     name: str
-    call_count: int
-    blocks_covered: float
+    call_count: Optional[int]
+    blocks_covered: Optional[float]
 
 
 class _FunctionSpecializationNameLine(NamedTuple):
@@ -301,8 +304,6 @@ def parse_coverage(
                 persistent_states,
             )
             tokenized_lines.append((parsed_line, raw_line))
-            if activate_trace_logging:
-                LOGGER.trace("Parsed line: %s", parsed_line)
         except Exception as ex:  # pylint: disable=broad-except
             lines_with_errors.append((raw_line, ex))
 
@@ -349,11 +350,13 @@ def parse_coverage(
         )
 
     filecov = FileCoverage(data_filename, filename=filename)
-    state = _ParserState()
-    if use_existing_files and missing_function_lines:
-        state = state._replace(function_name="<unknown>")
+    state = _ParserState.new(
+        function_name=UNKNOWN_FUNCTION_NAME,
+    )
     for line, raw_line in tokenized_lines:
         try:
+            if activate_trace_logging:
+                LOGGER.trace("Processing line: %s", line)
             state = _gather_coverage_from_line(
                 state,
                 line,
@@ -362,7 +365,7 @@ def parse_coverage(
             )
         except Exception as ex:  # pylint: disable=broad-except
             lines_with_errors.append((raw_line, ex))
-            state = _ParserState(is_recovering=True)
+            state = _ParserState.new(is_recovering=True)
 
     # Clean up the final state. This shouldn't happen,
     # but the last line could theoretically contain pending function lines
@@ -373,9 +376,26 @@ def parse_coverage(
             MergeOptions(func_opts=FUNCTION_MAX_LINE_MERGE_OPTIONS),
             mangled_name=name,
             demangled_name=None,
-            lineno=(state.linecov_list[-1].lineno + 1) if state.linecov_list else 0,
+            lineno=(state.last_linecov.lineno + 1) if state.last_linecov else 0,
             count=count,
             blocks=blocks,
+        )
+
+    unknown_function_linecovs = [
+        linecov
+        for linecov in filecov.linecov()
+        if linecov.function_name == UNKNOWN_FUNCTION_NAME
+    ]
+    if unknown_function_linecovs:
+        filecov.insert_function_coverage(
+            str(data_filename),
+            MergeOptions(func_opts=FUNCTION_MAX_LINE_MERGE_OPTIONS),
+            mangled_name=UNKNOWN_FUNCTION_NAME,
+            demangled_name=None,
+            lineno=unknown_function_linecovs[0].lineno,
+            count=None,
+            blocks=None,
+            excluded=True,
         )
 
     _report_lines_with_errors(
@@ -402,7 +422,7 @@ def _reconstruct_source_code(tokens: Iterable[_Line]) -> list[str]:
 class _ParserState(NamedTuple):
     """State information while parsing gcov lines.
 
-    >>> state = _ParserState()
+    >>> state = _ParserState.new()
     >>> state.previous_state is None
     True
     >>> state = state._replace(linecov_list=["x"])
@@ -428,20 +448,35 @@ class _ParserState(NamedTuple):
     >>> restored_state.restore_state()
     Traceback (most recent call last):
     ...
-    RuntimeError: Sanity check failed, previous_state of _ParserState is None.
+    gcovr.exceptions.SanityCheckError: Previous_state of _ParserState is None.
     """
 
     deferred_functions: list[_FunctionLine] = []
     function_name: Optional[str] = None
     function_specialization: bool = False
+    last_linecov: Optional[LineCoverage] = None
     linecov_list: list[LineCoverage] = []
     block_id: Optional[int] = None
     is_recovering: bool = False
     previous_state: Optional["_ParserState"] = None
 
+    @classmethod
+    def new(cls, **kwargs: Any) -> "_ParserState":
+        """Return a newly initialized parser state object."""
+        # We need to overwrite the list values because the default values are shared
+        # and changed.
+        return (
+            cls()
+            ._replace(
+                deferred_functions=[],
+                linecov_list=[],
+            )
+            ._replace(**kwargs)
+        )
+
     def save_state(self) -> "_ParserState":
         """Save the current parser state and return a new one with the current line coverage list."""
-        return _ParserState(
+        return _ParserState.new(
             previous_state=self,
             linecov_list=self.linecov_list,
         )
@@ -449,9 +484,7 @@ class _ParserState(NamedTuple):
     def restore_state(self) -> "_ParserState":
         """Restore the previous parser state with the current line coverage list."""
         if self.previous_state is None:
-            raise RuntimeError(
-                f"Sanity check failed, previous_state of {type(self).__name__} is None."
-            )
+            raise SanityCheckError(f"Previous_state of {type(self).__name__} is None.")
         return self.previous_state._replace(linecov_list=self.linecov_list)
 
 
@@ -475,11 +508,14 @@ def _gather_coverage_from_line(
     if isinstance(line, _SourceLine):
         raw_count, lineno, source_code, extra_info = line
 
+        is_noncode = extra_info & _ExtraInfo.NONCODE
+        if is_noncode:
+            return state  # skip noncode lines
+
         # handle deferred functions
         if state.deferred_functions:
             for function in state.deferred_functions:
                 name, count, blocks = function
-
                 filecov.insert_function_coverage(
                     filecov.data_sources,
                     MergeOptions(func_opts=FUNCTION_MAX_LINE_MERGE_OPTIONS),
@@ -489,62 +525,40 @@ def _gather_coverage_from_line(
                     count=count,
                     blocks=blocks,
                 )
+            state.deferred_functions.clear()
 
-            # If a specialized function is following a normal one the overall lines were added
-            # already to the previous function and we need to remove them again.
-            if (
-                state.function_specialization
-                and state.linecov_list
-                and state.linecov_list[0].lineno
-                < lineno
-                <= state.linecov_list[-1].lineno
-            ):
-                to_remove = [
-                    linecov
-                    for linecov in state.linecov_list
-                    if linecov.lineno >= lineno
-                ]
-                if len(to_remove) == 1:
-                    if activate_trace_logging:
-                        LOGGER.trace(
-                            "Removing line coverage for line %s from previous function.",
-                            to_remove[0].lineno,
-                        )
-                else:
-                    if activate_trace_logging:
-                        LOGGER.trace(
-                            "Removing line coverage for line %s to line %s from previous function.",
-                            to_remove[0].lineno,
-                            to_remove[-1].lineno,
-                        )
-                for linecov in to_remove:
-                    filecov.remove_line_coverage(linecov)
-            state = state._replace(
-                deferred_functions=[],
-                linecov_list=[],
-            )
+        # If a specialized function is following a normal one the overall lines were added
+        # already to the previous function and we need to remove them again.
+        if (
+            state.function_specialization
+            and state.linecov_list
+            and state.linecov_list[0].lineno <= lineno <= state.linecov_list[-1].lineno
+        ):
+            to_remove = [
+                linecov for linecov in state.linecov_list if linecov.lineno >= lineno
+            ]
 
-        is_noncode = extra_info & _ExtraInfo.NONCODE
-        if not is_noncode:
-            # This occurs if a file starts with a function with specializations.
-            # In this we get the overall line coverage of all specializations which we must ignore.
-            if state.function_name is None:
+            for linecov in to_remove:
                 if activate_trace_logging:
                     LOGGER.trace(
-                        "%s:%s: Ignoring line coverage because of missing function name: %s",
-                        filecov.location,
-                        lineno,
-                        source_code,
+                        "%s: Removing line coverage of function %s.",
+                        linecov.location,
+                        linecov.report_function_name,
                     )
-            else:
-                linecov = filecov.insert_line_coverage(
-                    filecov.data_sources,
-                    lineno=lineno,
-                    count=raw_count,
-                    function_name=state.function_name,
-                    md5=get_md5_hexdigest(source_code.encode("UTF-8")),
-                )
-                state.linecov_list.append(linecov)
+                filecov.remove_line_coverage(linecov)
+
+            state.linecov_list.clear()
+
+        linecov = filecov.insert_line_coverage(
+            filecov.data_sources,
+            lineno=lineno,
+            count=raw_count,
+            function_name=state.function_name,
+            md5=get_md5_hexdigest(source_code.encode("UTF-8")),
+        )
+        state = state._replace(last_linecov=linecov)
+        if not state.function_specialization:
+            state.linecov_list.append(linecov)
 
         return state
 
@@ -554,12 +568,11 @@ def _gather_coverage_from_line(
     elif isinstance(line, _FunctionLine):
         # Defer handling of the function tag until the next source line.
         # This is important to get correct line number information.
-        if len(state.deferred_functions):
-            if activate_trace_logging:
-                LOGGER.trace(
-                    "Several function definitions for the next coverage lines, function %s will not contain line coverage.",
-                    state.deferred_functions[-1].name,
-                )
+        if state.deferred_functions and activate_trace_logging:
+            LOGGER.trace(
+                "Several function definitions for the next coverage lines, function %s will not contain line coverage.",
+                state.deferred_functions[-1].name,
+            )
         return state._replace(
             deferred_functions=[*state.deferred_functions, line],
             function_name=line.name,
@@ -581,8 +594,8 @@ def _gather_coverage_from_line(
     elif isinstance(line, _BranchLine):
         branchno, hits, annotation = line
 
-        if state.linecov_list:
-            state.linecov_list[-1].insert_branch_coverage(
+        if state.last_linecov is not None:
+            state.last_linecov.insert_branch_coverage(
                 filecov.data_sources,
                 branchno=branchno,
                 count=hits,
@@ -598,8 +611,8 @@ def _gather_coverage_from_line(
         callno, returned = line
 
         # linecov won't exist if it was considered noncode
-        if state.linecov_list:
-            state.linecov_list[-1].insert_call_coverage(
+        if state.last_linecov is not None:
+            state.last_linecov.insert_call_coverage(
                 filecov.data_sources,
                 callno=callno,
                 source_block_id=state.block_id,  # type: ignore [arg-type]