+++ CHANGELOG.rst
@@ -47,6 +47,7 @@ Bug fixes and small improvements:
 - Add warning if function lines are missing in external generated GCOV files. (:issue:`1199`)
 - Extend support for zipped reports. If last suffix is ``.xz`` then LZMA is used to compress the report. (:issue:`1201`)
 - Fix function exclusion in report generation. (:issue:`1206`)
+- Do not ignore lines without functions, e.g. from inlined code. (:issue:`1204`)
 
 Documentation: